Refactor mod menu
