+++ MOD.Scripts.UI/MODMenuModuleInterface.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MOD.Scripts.UI
+{
+	interface MODMenuModuleInterface
+	{
+		void OnGUI();
+		void OnBeforeMenuVisible();
+	}
+}
